+++ data/src/test/java/com/london/data/repository/ActorRepositoryImplTest.kt
@@ -5,13 +5,13 @@ import com.london.data.mapper.details.actor.toEntity
 import com.london.data.mapper.details.tvshow.toCastEntity
 import com.london.data.remote.exception.NetworkException
 import com.london.data.remote.model.ApiResponse
-import com.london.data.remote.model.details.actor.model.ActorDetailsResponse
-import com.london.data.remote.model.details.actor.model.actorimage.ActorImageResponse
-import com.london.data.remote.model.details.movie.model.moviecast.MovieActor
-import com.london.data.remote.model.details.movie.model.moviecast.MovieCastResponse
-import com.london.data.remote.model.details.tvshow.model.Role
-import com.london.data.remote.model.details.tvshow.model.TvShowCastMember
-import com.london.data.remote.model.details.tvshow.model.TvShowCastRemoteResponse
+import com.london.data.remote.model.details.actor.ActorDetailsResponse
+import com.london.data.remote.model.details.actor.actorimage.ActorImageResponse
+import com.london.data.remote.model.details.movie.moviecast.MovieActor
+import com.london.data.remote.model.details.movie.moviecast.MovieCastResponse
+import com.london.data.remote.model.details.tvshow.Role
+import com.london.data.remote.model.details.tvshow.TvShowCastMember
+import com.london.data.remote.model.details.tvshow.TvShowCastRemoteResponse
 import com.london.data.remote.model.home.trending.TrendingResponse
 import com.london.data.remote.source.actor.ActorRemoteDataSource
 import com.london.data.repository.actor.ActorRepositoryImpl
