+++ data/src/test/java/com/london/data/repository/MovieRepositoryImplTest.kt
@@ -10,14 +10,14 @@ import com.london.data.mapper.details.movie.toEntity
 import com.london.data.mapper.search.toAuthorDetails
 import com.london.data.remote.exception.NetworkException
 import com.london.data.remote.model.ApiResponse
-import com.london.data.remote.model.details.movie.model.moviedetails.GenreRemote
-import com.london.data.remote.model.details.movie.model.moviedetails.MovieDetailsResponse
-import com.london.data.remote.model.details.movie.model.moviedetails.ProductionCompanyRemote
-import com.london.data.remote.model.details.movie.model.moviedetails.ProductionCountryRemote
-import com.london.data.remote.model.details.movie.model.moviedetails.RemoteCollectionDetails
-import com.london.data.remote.model.details.movie.model.moviedetails.SpokenLanguageRemote
-import com.london.data.remote.model.details.movie.model.movieimages.MovieImagesResponse
-import com.london.data.remote.model.details.movie.model.movieimages.Poster
+import com.london.data.remote.model.details.movie.moviedetails.GenreRemote
+import com.london.data.remote.model.details.movie.moviedetails.MovieDetailsResponse
+import com.london.data.remote.model.details.movie.moviedetails.ProductionCompanyRemote
+import com.london.data.remote.model.details.movie.moviedetails.ProductionCountryRemote
+import com.london.data.remote.model.details.movie.moviedetails.RemoteCollectionDetails
+import com.london.data.remote.model.details.movie.moviedetails.SpokenLanguageRemote
+import com.london.data.remote.model.details.movie.movieimages.MovieImagesResponse
+import com.london.data.remote.model.details.movie.movieimages.Poster
 import com.london.data.remote.model.details.rating.AccountStatesResponse
 import com.london.data.remote.model.details.rating.RatingRemoteResponse
 import com.london.data.remote.model.home.popular.PopularMovieResponse
